+++ core/execution/market.go
@@ -2558,7 +2558,10 @@ func (m *Market) amendOrder(
 
 		// Update the existing message in place before we cancel it
 		if foundOnBook {
-			m.orderAmendInPlace(existingOrder, amendedOrder)
+			// Do not amend in place, the amened could be something
+			// not support for an amend in place, and not pass
+			// the validation of the order book
+			// m.orderAmendInPlace(existingOrder, amendedOrder)
 			cancellation, err := m.matching.CancelOrder(amendedOrder)
 			if cancellation == nil || err != nil {
 				m.log.Panic("Failure to cancel order from matching engine",
@@ -2566,11 +2569,12 @@ func (m *Market) amendOrder(
 					logging.String("order-id", amendedOrder.ID),
 					logging.String("market", m.mkt.ID),
 					logging.Error(err))
-				return nil, nil, err
 			}
 
-			_ = m.position.UnregisterOrder(ctx, cancellation.Order)
-			amendedOrder = cancellation.Order
+			// unregister the existing order
+			// _ = m.position.UnregisterOrder(ctx, cancellation.Order)
+			_ = m.position.UnregisterOrder(ctx, existingOrder)
+			// amendedOrder = cancellation.Order
 		}
 
 		// Update the order in our stores (will be marked as cancelled)